Fixed bug in printing CSR instructions with immediate: Immediate value was being printed with a leading x.
To be compatible with spike tracer, we currently skip printing changed
CSRs.

Each CSR is now associated with a mask. A zero bit in the mask
preserved the corresponding bit in the CSR value. A one bit makes the
corresponding bit in the CSR value writable.

The Csr class fields are now private. All access is through methods.

The mtval CSR is now associated with a mask of ~2 (bit 2 is not writable).

Now loading hex file if --hex option is specified.

Fixed bug in hex-file loader: We were dropping the last value in a
line.