When changing load or data bit or both in the debug csr (dcsr), we turn off the load bit to disallow the load/data combination.
We no longer check privilege mode when poking/peeking CSRs.

We no longer check data/instruction attribute when peeking instruction/data.

We no longer check data/instruction attribute when poking instruction/data.

Fixed bug in instruction tracing: We were reporting the data1
component of a trigger regardless of which component was actually
modified.

Trigger registers now writeable in M mode only of the dmode bit of
data1 of selected trigger is is set to 0.

The dmode bit of data1 component of a trigger is now writable only in
debug mode.